Very basic simulation.
Includes concepts of features (use cases) and blockers.

A given user wants to use the product if it implements any of their use
cases, but can't use it if it does not implement all their blockers.

However, if it implements one of their use cases and they can't adopt
it due to a blocker, they still have a chance to learn about the
product (ie. awareness). This can lead to faster adoption later once
the blocker(s) are resolved.