Change function for assigning wants/needs to users.
Instead of a fixed N needs or wants per user, use a fixed
probability of each need or want. This allows users to have from 0..M
needs or wants, but still an average of N. This is more like reality
and also gives smoother curves.

With this, we can remove the weird logic to pre-supply the first 10
needs so that we don't have so much zero time at the beginning. The
numbers are still very low at first, but that's because the simulation
is similar to reality (if you only clear one or two roadblocks,
vanishingly few people will be able to use the product).